+++ blt/urls.py
@@ -53,7 +53,11 @@ from website.views import (
     github_callback,
     google_callback,
     facebook_callback,
-    sponsor_view
+    sponsor_view,
+    like_issue2,
+    flag_issue2,
+    subscribe_to_domains,
+    IssueView2
 )
 from website.api.views import (
     IssueViewSet,
@@ -70,12 +74,6 @@ from website.api.views import (
     InviteFriendApiViewset
 )
 from company.views import ShowBughuntView
-from website.alternative_views import (
-    like_issue2,
-    flag_issue2,
-    subscribe_to_domains,
-    IssueView2
-)
 
 from blt import settings
 from rest_framework import permissions, routers
